agent_afford_harness/requirements.txt
numpy>=1.20
Pillow>=9.0
PyYAML>=6.0
opencv-python>=4.5
transformers>=4.45
openai>=1.0
